Fix meal history and dashboard issues

+++ src/main/java/com/santosh/nutrition_analyzer/dto/food/DashboardResponse.java
@@ -21,6 +21,6 @@ public class DashboardResponse {
 
     private Double totalFiber;
 
-    private int averageHealthScore;
+    private Double averageHealthScore;
 
 }
